Fix basic warnings

+++ apps/solver/src/Scanner.hpp
@@ -14,7 +14,7 @@ class scanner {
   std::string m_source;
   std::vector<token> m_tokens;
   int m_start = 0;
-  int m_current = 0;
+  size_t m_current = 0;
   int m_line = 1;
 
 public:
